Remove uses of Time.now in P2p_acl

+++ src/lib_p2p/p2p.ml
@@ -43,7 +43,7 @@ type limits = {
 
   connection_timeout : float ;
   authentication_timeout : float ;
-  greylist_timeout : float ;
+  greylist_timeout : int ;
 
   min_connections : int ;
   expected_connections : int ;
